Fixed valuepack to compile with GCC on Linux
Added device_nop, which doesn't actually create a device.
This allows interop with frameworks which create a context
and window themselves.

Added namespace to avoid Font class name clash (TODO put
everything else in namespace)

Changed CharRange's read_float to cope with NaNs

+++ include/core/indexbuffer.h
@@ -10,21 +10,21 @@ class IndexBuffer : public Shared
 {
 public:
 	typedef SharedPtr< IndexBuffer > Ptr;
-	explicit IndexBuffer( int count = 0, unsigned short *indices = 0, bool dynamic = false ) : m_dynamic( dynamic ), m_gl_buffer( 0 )
+	explicit IndexBuffer( int count = 0, unsigned int *indices = 0, bool dynamic = false ) : m_dynamic( dynamic ), m_gl_buffer( 0 )
 	{
 		if( count ) m_indices.resize( count );
 		if( count && indices ) std::copy( indices, indices + count, m_indices.begin() );
 	}
 
-	IndexBuffer &add( unsigned short index ) {m_indices.push_back( index ); return *this;}
+	IndexBuffer &add( unsigned int index ) {m_indices.push_back( index ); return *this;}
 
-	unsigned short *indices();
+	unsigned int *indices();
 
 	int count() const {return m_indices.size();}
 	void clear() {m_indices.clear();}
 
 private:
-	std::vector< unsigned short > m_indices;
+	std::vector< unsigned int > m_indices;
 	bool m_dynamic;
 
 	unsigned int m_gl_buffer;
